+++ src/de/mod.rs
@@ -3,12 +3,13 @@ use std::io::Read;
 
 use self::read::{BincodeRead, IoReader, SliceReader};
 use byteorder::ReadBytesExt;
-use internal::SizeLimit;
+use config::SizeLimit;
 use serde;
 use serde::de::Error as DeError;
 use serde::de::IntoDeserializer;
 use {Error, ErrorKind, Result};
 
+/// Specialized ways to read data into bincode.
 pub mod read;
 
 /// A Deserializer that reads bytes from a buffer.
@@ -30,12 +31,12 @@ pub struct Deserializer<R, O: Options> {
 }
 
 impl<'de, R: BincodeRead<'de>, O: Options> Deserializer<R, O> {
-    /// Creates a new Deserializer with a given `Read`er and a size_limit.
+    /// Creates a new Deserializer with a given `Read`er and options.
     pub fn new<IR: Read>(r: IR, options: O) -> Deserializer<IoReader<IR>, O> {
         Deserializer { reader: IoReader::new(r), options }
     }
     
-    ///
+    /// Creates a new Deserializer that will read from the given slice.
     pub fn from_slice(slice: &'de [u8], options: O) -> Deserializer<SliceReader<'de>, O> {
         Deserializer {
             reader: SliceReader::new(slice),
@@ -43,7 +44,7 @@ impl<'de, R: BincodeRead<'de>, O: Options> Deserializer<R, O> {
         }
     }
 
-    ///
+    /// Creates a new Deserializer with the given `BincodeRead`er
     pub fn with_bincode_read(r: R, options: O) -> Deserializer<R, O> {
         Deserializer {
             reader: r,
